+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 - Shortcut for calling & validation. `#738`_
 - New hook to pre-commit, rstcheck, as well as updates to documentation based on rstcheck `#734`_
 - New check for maximum response time and corresponding CLI option ``--max-response-time`` `#716`_
+- New field with information about executed checks in cassettes. `#702`_
 
 **Fixed**
 
@@ -1385,6 +1386,7 @@ Deprecated
 .. _#716: https://github.com/schemathesis/schemathesis/issues/716
 .. _#708: https://github.com/schemathesis/schemathesis/issues/708
 .. _#705: https://github.com/schemathesis/schemathesis/issues/705
+.. _#702: https://github.com/schemathesis/schemathesis/issues/702
 .. _#692: https://github.com/schemathesis/schemathesis/issues/692
 .. _#686: https://github.com/schemathesis/schemathesis/issues/686
 .. _#684: https://github.com/schemathesis/schemathesis/issues/684